ResumeATS/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
openai
python-dotenv
